Migrated langroles and removed old admin file
